requirements.txt
apache-airflow-providers-mysql==5.3.0
apache-airflow-providers-apache-spark==4.1.0
pyspark==3.5.0
mysqlclient==2.2.1           # Cliente MySQL para Python
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
xgboost==1.7.5                # Opcional: Si usas Pandas en PySpark
mysql-connector-python==8.1.0
pymysql==1.0.2